requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blinker==1.8.2
boto3==1.35.46
botocore==1.35.46
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
compressed-tensors==0.6.0
crashtest==0.4.1
dataclasses-json==0.6.7
datasets==3.1.0
debugpy==1.8.8
decorator==5.1.1
deepeval==1.4.8
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docx2txt==0.8
dulwich==0.21.7
einops==0.8.0
Events==0.5
execnet==2.1.1
executing==2.1.0
fastapi==0.115.5
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
flow-judge==0.1.2
frozenlist==1.5.0
fsspec==2024.6.1
gguf==0.10.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.63.2
h11==0.14.0
hf_transfer==0.1.8
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.7
importlib-metadata==7.0.0
iniconfig==2.0.0
installer==0.7.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keyring==24.3.1
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
langsmith==0.1.139
lark==1.2.2
llama-cloud==0.1.5
llama-index==0.11.23
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.23
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.5.0
llama-index-legacy==0.9.48.post4
llama-index-llms-ollama==0.3.6
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.3.0
llama-index-readers-llama-parse==0.3.0
llama-index-storage-docstore-redis==0.2.0
llama-index-storage-index-store-redis==0.3.0
llama-index-storage-kvstore-redis==0.2.0
llama-index-vector-stores-opensearch==0.4.1
llama-parse==0.5.14
llvmlite==0.43.0
lm-format-enforcer==0.10.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.4.4
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-ml-py==12.560.30
ollama==0.3.3
openai==1.53.1
opencv-contrib-python-headless==4.10.0.84
opencv-python-headless==4.10.0.84
opensearch==0.9.2
opensearch-py==2.7.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.11
outlines==0.0.46
packaging==24.1
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.2
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
portalocker==2.10.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.0.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pypdf==5.1.0
pyproject_hooks==1.2.0
pysbd==0.3.4
pytest==8.3.3
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.2.3
RapidFuzz==3.10.1
ray==2.39.0
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
sentencepiece==0.2.0
sentry-sdk==2.18.0
setuptools==75.5.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.41.2
striprtf==0.0.26
structlog==24.4.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
tomlkit==0.13.2
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.2
trove-classifiers==2024.10.21.16
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.27.1
vllm==0.6.3.post1
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wrapt==1.16.0
xattr==1.1.0
xxhash==3.5.0
yarl==1.17.1
zipp==3.20.2
